The Real Sharks in the Tank
Using Power BI, I have created a dashboard on Shark Tank India Season 1,2 & 3. you will be able to find the details of total pitches, total invested, no. of companies and many more.
The data was collected from different sources and cleaning of data was the major role in this process.